- connection area defined by the decorator in DOM - overlap detection and minimum shift handled in ModelEditorView - connection points are defined in DOM - connection area divided between the connections equally - connections segmentpoints are snapped to neighboruring ports - selection menu v0.1 impleemnted

+++ nodeserver/src/client/js/GraphViz/GraphVizControl.js
@@ -215,8 +215,10 @@ define(['logManager',
 
             for (i = 0; i < childrenDiff.length; i += 1) {
                 childId = childrenDiff[i];
-                this._graphVizView.deleteObject(this._components[childId].componentInstance, this._components[objectId].componentInstance);
-                delete this._components[childId];
+                if (this._components[childId] && this._components[childId].componentInstance) {
+                    this._graphVizView.deleteObject(this._components[childId].componentInstance, this._components[objectId].componentInstance);
+                    delete this._components[childId];
+                }
             }
 
             //Handle children addition
